bumping postcards from 300 chars to 100 words
